+++ api/V0ComposeKillTask.go
@@ -12,16 +12,16 @@ import (
 // V0ComposeKillTask will kill a task from a service from a specific project
 // example:
 // curl -X DELETE http://user:password@127.0.0.1:10000/api/compose/v0/{projectname}/{servicename}/{taskid}
-func V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
+func (e *API) V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
-	auth := CheckAuth(r, w)
+	auth := e.CheckAuth(r, w)
 
 	if vars == nil || !auth {
 		return
 	}
 
 	logrus.Debug("HTTP DELETE V0ComposeKillTask")
-	d := ErrorMessage(2, "V0ComposeKillTask", "nok")
+	d := e.ErrorMessage(2, "V0ComposeKillTask", "nok")
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")
 
@@ -34,7 +34,7 @@ func V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
 	servicename := vars["servicename"]
 	taskID := vars["taskid"]
 
-	key := GetRedisKey(config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
+	key := e.GetRedisKey(e.Config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
 
 	var task mesosutil.Command
 	json.Unmarshal([]byte(key), &task)
@@ -42,14 +42,14 @@ func V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
 	if task.TaskID == taskID {
 		err := mesosutil.Kill(task.TaskID, task.Agent)
 		if err != nil {
-			d = ErrorMessage(2, "V0ComposeKillTask", err.Error())
+			d = e.ErrorMessage(2, "V0ComposeKillTask", err.Error())
 			logrus.Error("V0ComposeKillTask Error during kill: ", err)
 			w.Write(d)
 			return
 		}
 
-		logrus.Debug("V0ComposeKillTask: " + config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
-		d = ErrorMessage(0, "V0ComposeKillTask", "ok")
+		logrus.Debug("V0ComposeKillTask: " + e.Config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
+		d = e.ErrorMessage(0, "V0ComposeKillTask", "ok")
 	}
 	w.Write(d)
 }